Fix silent SDP media attribute counter overflow (#2374)
MAX_SDP_ATTRIBUTES_COUNT is 256 but SdpMediaDescription counted media attributes in a UINT8, so the bounds check in parseMediaAttributes() could never fail. Once 256 attributes were stored the counter wrapped to 0 and later attributes overwrote sdpAttributes[0] onward, where browser offers keep the ICE and DTLS parameters. deserializeSessionDescription() still returned STATUS_SUCCESS.

Widen the counter to UINT16 so the existing check works, and widen the matching loop counter in setPayloadTypesFromOffer(), which would otherwise never terminate once a media section holds 256 attributes.

+++ src/source/PeerConnection/SessionDescription.c
@@ -193,7 +193,7 @@ STATUS setPayloadTypesFromOffer(PHashTable codecTable, PHashTable rtxTable, PSes
     ENTERS();
     STATUS retStatus = STATUS_SUCCESS;
     PSdpMediaDescription pMediaDescription = NULL;
-    UINT8 currentAttribute;
+    UINT16 currentAttribute;
     UINT16 currentMedia;
     PCHAR attributeValue, end;
     UINT64 parsedPayloadType, hashmapPayloadType, fmtpVal, aptVal;
